+++ lib/cr_comment/src/due.ml
@@ -19,7 +19,8 @@
 (*  <http://www.gnu.org/licenses/> and <https://spdx.org>, respectively.        *)
 (********************************************************************************)
 
-module Due = Due
-module Kind = Kind
-module Filter = Filter
-include Cr_comment0
+type t =
+  | Now
+  | Soon
+  | Someday
+[@@deriving compare, equal, enumerate, sexp_of]