+++ lib/crs_cli/src/common_helpers.mli
@@ -34,3 +34,9 @@ val relativize
   -> cwd:Absolute_path.t
   -> path:Fpath.t
   -> Vcs.Path_in_repo.t
+
+(** Allows to use flags to supply CR filters. For example: [--xcrs] or [--now].
+    If no filter flags are supplied, this returns [`Default], which may be
+    interpreted differently depending on the context. Each command is
+    responsible for documenting the meaning of its filter flags. *)
+val filters : [ `Default | `Supplied of Cr_comment.Filter.t list ] Command.Arg.t